MOVED - Interceptions.jsx into tabs/ops folder
